:recycle: Refactor: 밈 상세 내 버튼들 파일 이동
